requirements.txt
aniso8601==4.1.0
coverage==4.5.2
Django==2.1.5
django-cors-headers==2.4.0
django-graphql-jwt==0.2.0
django-nose==1.4.6
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
nose==1.3.7
promise==2.2.1
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.9
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
